test: Split integration and binding scaling e2e tests
